refactor: 티켓팅 생성 api 분리
