+++ ticatch-front/lib/axiosClient.ts
@@ -13,7 +13,10 @@ export const axiosClient = axios.create({
 
 axiosClient.interceptors.request.use(
   (config) => {
-    const accessToken = localStorage.getItem('accessToken');
+    // 임시 엑세스토큰
+    const accessToken =
+      'eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJib2F6NDFAbmF2ZXIuY29tIiwiYXV0aCI6IlVTRVIiLCJleHAiOjE3Mzc3MDc4OTR9.K1iWRehuFVXPuDzOq-DIsflIujEGwVg1tYUCSa7VRTONDVsdXVAOf70U7GcRHtUPFicuI10QpnyOCJWiLCfa1Q';
+    // const accessToken = localStorage.getItem('accessToken');
 
     // 토큰 만료
     if (accessToken && isTokenExpired(accessToken)) {
